deploy annotations in standalone mode
