--- deploy/Dockerfile
@@ -1,27 +1,0 @@
-FROM ubuntu:16.04
-
-ENV TERM=xterm
-
-ARG PROJ
-
-RUN \
-  apt-get update && \
-  apt-get install -y supervisor && \
-  rm -rf /var/lib/apt/lists/* && \
-  sed -i 's/^\(\[supervisord\]\)$/\1\nnodaemon=true/' /etc/supervisor/supervisord.conf
-
-VOLUME ["/etc/supervisor/conf.d"]
-
-ENV GOPATH /go
-ENV PATH $GOPATH/bin:$PATH
-
-RUN mkdir -p "$GOPATH/src" "$GOPATH/bin" && chmod -R 777 "$GOPATH"
-
-COPY bin/$PROJ $GOPATH/bin
-RUN ln -s $GOPATH/bin/$PROJ $GOPATH/bin/go_app 
-
-COPY supervisord.conf /etc/supervisor
-
-WORKDIR /etc/supervisor/conf.d
-
-ENTRYPOINT ["supervisord", "-c", "/etc/supervisor/supervisord.conf"]
