+++ deploy/Dockerfile-standalone
@@ -1,0 +1,20 @@
+FROM ubuntu:16.04
+
+ENV TERM=xterm
+
+ARG PROJ
+
+RUN \
+  apt-get update
+
+ENV GOPATH /go
+ENV PATH $GOPATH/bin:$PATH
+
+RUN mkdir -p "$GOPATH/src" "$GOPATH/bin" && chmod -R 777 "$GOPATH"
+
+COPY bin/$PROJ $GOPATH/bin
+RUN ln -s $GOPATH/bin/$PROJ $GOPATH/bin/go_app 
+
+WORKDIR $GOPATH/bin
+
+ENTRYPOINT ["go_app"]
