+++ deploy/build.sh
@@ -2,20 +2,22 @@
 
 set -o errexit
 
-if [ "$#" -lt 2 ]; then
-  echo "usage: $0 project tag [push]"
+if [ "$#" -lt 3 ]; then
+  echo "usage: $0 {project} {tag} {supervisord | standalone} [push]"
   exit 1
 fi
 
 proj=$1
 tag=$2
+mode=$3
 docker_proj=go_watchman_$proj:$tag
 
 set -x
 
-go install github.com/Sotera/go_watchman/$proj
-cp $GOPATH/bin/$proj ./bin
-docker build -t sotera/$docker_proj --build-arg PROJ=$proj .
-if [ "$3" == "push" ]; then
+go test ../$proj
+# executables expected in 'cmd' dir
+go build -o bin/$proj github.com/Sotera/go_watchman/$proj/cmd
+docker build -f Dockerfile-$mode -t sotera/$docker_proj --build-arg PROJ=$proj .
+if [ "$4" == "push" ]; then
   docker push sotera/$docker_proj
 fi
